backend/packager/requirements.txt
pandas
scikit-learn
joblib
kafka-python
os
psycopg2
python-dotenv